+++ report.go
@@ -18,8 +18,8 @@ func (me *Report) AddEntries(v ...Entry) {
 	me.entries = append(me.entries, v...)
 }
 
-// Save saves the report as markdown to the given filename.
-func (me *Report) Save(filename string) error {
+// SaveAs saves the report as markdown to the given filename.
+func (me *Report) SaveAs(filename string) error {
 	fh, err := os.Create(filename)
 	if err != nil {
 		return err