+++ changelog.md
@@ -6,6 +6,7 @@ and this project adheres to http://semver.org/spec/v2.0.0.html.
 
 ## [unreleased]
 
+- Remove MustSetVerifiedNow
 - Add Entry.IsLevel
 
 ## [0.7.1] - 2021-03-03