+++ changelog.md
@@ -4,9 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on http://keepachangelog.com/en/1.0.0/
 and this project adheres to http://semver.org/spec/v2.0.0.html.
 
-## [unreleased]
+## [0.8.0] - 2021-03-05
 
-- Replace Editor.TidyExport with WriteTo implementing io.WriterTo interface
+- Replace Editor.TidyExport with WriteTo implementing io.WriterTo
 - Remove Editor.Export
 - Rename Editor.Save to Editor.SaveAs
 - Remove MustSetVerifiedNow